+++ openquake/hazardlib/gsim/can15/western.py
@@ -69,22 +69,23 @@ class WesternCan15Mid(BooreAtkinson2011):
     sgn = 0
 
     def get_mean_and_stddevs(self, sites, rup, dists, imt, stddev_types):
-        """ """
         # distances
-        distsl = copy.copy(dists)
-        distsl.rjb, distsl.rrup = \
-            get_equivalent_distances_west(rup.mag, dists.repi)
+        if self.REQUIRES_DISTANCES == {'repi'}:
+            # convert from repi to rjb
+            dists = copy.copy(dists)
+            dists.rjb, dists.rrup = get_equivalent_distances_west(
+                rup.mag, dists.repi)
         # get original values
-        mean, stddevs = super().get_mean_and_stddevs(sites, rup, distsl, imt,
+        mean, stddevs = super().get_mean_and_stddevs(sites, rup, dists, imt,
                                                      stddev_types)
         if self.sgn:
             # adjust mean values using the reccomended delta (see Atkinson and
             # Adams, 2013)
-            tmp = 0.1+0.0007*distsl.rjb
+            tmp = 0.1+0.0007*dists.rjb
             tmp = np.vstack((tmp, np.full_like(tmp, 0.3)))
             delta = np.log(10.**(np.amin(tmp, axis=0)))
             mean += self.sgn * delta
-        stds = [np.ones(len(distsl.rjb))*get_sigma(imt)]
+        stds = [np.ones(len(dists.rjb))*get_sigma(imt)]
         return mean, stds
 
 
@@ -93,4 +94,35 @@ class WesternCan15Low(WesternCan15Mid):
 
 
 class WesternCan15Upp(WesternCan15Mid):
+    sgn = +1
+
+
+class WesternCan15RjbMid(WesternCan15Mid):
+    """
+    Implements the Boore and Atkinson (2008) with adjustments proposed by
+    Boore and Atkinson (2011) and the modifications introduced for the
+    calculation of hazard for the fifth generation of Canada's hazard maps,
+    released in 2015.
+    """
+    #: GMPE not tested against independent implementation so raise
+    #: not verified warning
+    non_verified = True
+
+    #: Shear-wave velocity for reference soil conditions in [m s-1]
+    DEFINED_FOR_REFERENCE_VELOCITY = 760.
+
+    #: Standard deviation types supported
+    DEFINED_FOR_STANDARD_DEVIATION_TYPES = {StdDev.TOTAL}
+
+    #: Use rjb as the original BooreAtkinson
+    REQUIRES_DISTANCES = {'rjb'}
+
+    sgn = 0
+
+
+class WesternCan15RjbLow(WesternCan15RjbMid):
+    sgn = -1
+
+
+class WesternCan15RjbUpp(WesternCan15RjbMid):
     sgn = +1
