--- openquake/hazardlib/gsim/can15/western_rjb.py
@@ -1,70 +1,0 @@
-#
-# Copyright (C) 2014-2021 GEM Foundation
-#
-# OpenQuake is free software: you can redistribute it and/or modify it
-# under the terms of the GNU Affero General Public License as published
-# by the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# OpenQuake is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-#
-# You should have received a copy of the GNU Affero General Public License
-# along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
-
-
-"""
-:module:`openquake.hazardlib.gsim.can15.western` implements
-:class:`WesternCan15RjbMid`, :class:`WesternCan15RjbLow`,
-:class:`WesternCan15RjbUpp`
-"""
-
-import numpy as np
-from openquake.hazardlib.gsim.can15.western import get_sigma
-from openquake.hazardlib.gsim.boore_atkinson_2011 import BooreAtkinson2011
-from openquake.hazardlib.const import StdDev
-
-
-class WesternCan15RjbMid(BooreAtkinson2011):
-    """
-    Implements the Boore and Atkinson (2008) with adjustments proposed by
-    Boore and Atkinson (2011) and the modifications introduced for the
-    calculation of hazard for the fifth generation of Canada's hazard maps,
-    released in 2015.
-    """
-    #: GMPE not tested against independent implementation so raise
-    #: not verified warning
-    non_verified = True
-
-    #: Shear-wave velocity for reference soil conditions in [m s-1]
-    DEFINED_FOR_REFERENCE_VELOCITY = 760.
-
-    #: Standard deviation types supported
-    DEFINED_FOR_STANDARD_DEVIATION_TYPES = {StdDev.TOTAL}
-
-    sgn = 0
-
-    def get_mean_and_stddevs(self, sites, rup, dists, imt, stddev_types):
-        """ """
-        # get original values
-        mean, stddevs = super().get_mean_and_stddevs(sites, rup, dists, imt,
-                                                     stddev_types)
-        if self.sgn:
-            # adjust mean values using the reccomended delta (see Atkinson and
-            # Adams, 2013)
-            tmp = 0.1+0.0007*dists.rjb
-            tmp = np.vstack((tmp, np.ones_like(tmp)*0.3))
-            delta = np.log(10.**(np.amin(tmp, axis=0)))
-            mean += self.sgn * delta
-        stds = [np.ones(len(dists.rjb))*get_sigma(imt)]
-        return mean, stds
-
-
-class WesternCan15RjbLow(WesternCan15RjbMid):
-    sgn = -1
-
-
-class WesternCan15RjbUpp(WesternCan15RjbMid):
-    sgn = +1
